設定ファイルテキストのプレビュー機能をコンポーネント化
