+++ src/lib/common.ts
@@ -1,5 +1,18 @@
+import { palworldServerSettings } from '$lib/palworld-server-serttings';
 import type { IPalworldServerSetting } from "$lib/types";
 import type { SelectOptionType } from "flowbite-svelte";
+
+/**
+ * 選択中のゲームバージョンの設定を取得
+ * @param selectedVersion 選択中のゲームバージョン
+ */
+export const selectedVersionSettings = (selectedVersion: string) => {
+    const settings = palworldServerSettings.filter(
+        (settings) => settings.version === selectedVersion
+    )[0].settings;
+
+    return settings;
+}
 
 /**
  * 複数の選択肢を取得
