+++ src/routes/+page.svelte
@@ -24,8 +24,20 @@
 	import IntSetting from '$lib/components/form/IntSetting.svelte';
 	import FloatSetting from '$lib/components/form/FloatSetting.svelte';
 	import BoolSetting from '$lib/components/form/BoolSetting.svelte';
+	import SettingTextPreview from '$lib/components/SettingTextPreview.svelte';
 
-
+	// フォーム
+	let formElement: Element;
+	// 無効化された項目を有効化
+	let forceEnableDisabledItems = false;
+	// 選択中のサーババージョン
+	let selectedServerVersion: string;
+	// 選択中のサーババージョンの設定
+	let selectedSettings: IPalworldServerSettings;
+	// フォームの入力値
+	let formValues: FormValues = {};
+	// サーバ設定ファイルテキスト
+	let settingText: string;
 
 	/**
 	 * 最新のサーババージョン
@@ -40,7 +52,7 @@
 	 * @param settingKey　設定キー
 	 */
 	const selectionItems = (settingKey: string): SelectOptionType<any>[] => {
-		const setting = selectedVersionSettings.filter((setting) => setting.key === settingKey)[0];
+		const setting = selectedSettings.filter((setting) => setting.key === settingKey)[0];
 		const items: SelectOptionType<any>[] = [];
 		setting.values.forEach((value) => {
 			const name = value.defalut ? `${value.value}(デフォルト)` : value.value.toString();
@@ -57,7 +69,7 @@
 	 * @param settingKey 設定キー
 	 */
 	const defaultSettingValue = (settingKey: string) => {
-		const setting = selectedVersionSettings.filter((setting) => setting.key === settingKey)[0];
+		const setting = selectedSettings.filter((setting) => setting.key === settingKey)[0];
 		const value = setting.values.filter((value) => value.defalut)[0].value;
 		return value;
 	};
@@ -72,27 +84,9 @@
 		return typed;
 	};
 
-
-
-	// フォーム
-	let formElement: Element;
-	// 無効化された項目を有効化
-	let forceEnableDisabledItems = false;
-	// 選択中のサーババージョン
-	let selectedServerVersion: string;
-	// 選択中のサーババージョンの設定
-	let selectedVersionSettings: IPalworldServerSettings;
-	// フォームの入力値
-	let formValues: FormValues = {};
-	$: {
-		selectedVersionSettings = palworldServerSettings.filter(
-			(settings) => settings.version === selectedServerVersion
-		)[0].settings;
-	}
-
 	// リアクティブされないことがあるので再定義
 	const initCurrentServerSettings = () => {
-		selectedVersionSettings = palworldServerSettings.filter(
+		selectedSettings = palworldServerSettings.filter(
 			(settings) => settings.version === selectedServerVersion
 		)[0].settings;
 	};
@@ -102,7 +96,7 @@
 	 */
 	const initFormValues = () => {
 		formValues = {};
-		selectedVersionSettings.forEach((setting) => {
+		selectedSettings.forEach((setting) => {
 			formValues[setting.key] = {
 				key: setting.key,
 				value: defaultSettingValue(setting.key),
@@ -166,7 +160,8 @@
 		const fileName = 'PalWorldSettings.ini';
 		const fileType = 'text/plain';
 
-		const text = createServerSettingFileText();
+		const text = settingText;
+		console.log(text);
 		const blob = new Blob([text], { type: fileType });
 
 		const tempA = document.createElement('a');
@@ -180,79 +175,6 @@
 		setTimeout(function () {
 			URL.revokeObjectURL(tempA.href);
 		}, 1500);
-	};
-
-	let serverSettingFileText = '';
-	$: {
-		// 変更検知のため読み出し
-		formValues;
-
-		serverSettingFileText = createServerSettingFileText();
-	}
-	/**
-	 * フォームの設定値からサーバの設定用テキストを作成(prefixを合わせた内容)
-	 */
-	const createServerSettingFileText = (): string => {
-		const settingFormat = selectedServerSettingFormat();
-		const formText = createFormSettingText();
-		const serverSettingText = settingFormat.replace(':GENERATE_SETTINGS:', formText);
-		return serverSettingText;
-	};
-
-	/**
-	 * フォームの設定値からサーバ設定用のテキストを作成(値部分)
-	 */
-	const createFormSettingText = (): string => {
-		const formatedValues: string[] = [];
-		// フォーム入力値は辞書型で順不同なので順番を保証する
-		selectedVersionSettings.forEach((value) => {
-			const formValue = formValues[value.key];
-			let formated = '';
-			// TODO: FormValueをinterfaceからclassにしてclass側でフォーマットする
-			switch (formValue.type) {
-				case 'planetext':
-					// 書式: key=planetext
-					formated = formValue.value.toString();
-					break;
-				case 'string':
-					// 書式: key="string"
-					formated = `\"${formValue.value.toString()}\"`;
-					break;
-				case 'int':
-					// 書式: key=123
-					const intNumber = Math.trunc(Number(formValue.value));
-					formated = intNumber.toString();
-					break;
-				case 'float':
-					// 書式: key=1.000000
-					const floatNumber = parseFloat(formValue.value as string);
-					formated = floatNumber.toFixed(6);
-					break;
-				case 'bool':
-					// 書式: key:True
-					const bool = formValue.value as boolean;
-					formated = bool ? 'True' : 'False';
-					break;
-				default:
-					console.warn('未対応');
-					break;
-			}
-
-			formatedValues.push(`${formValue.key}=${formated}`);
-		});
-
-		const text = formatedValues.join(',');
-		return text;
-	};
-
-	/**
-	 * 選択中のサーババージョンのサーバ設定書式
-	 */
-	const selectedServerSettingFormat = () => {
-		const format = palworldServerSettings.filter(
-			(settings) => settings.version === selectedServerVersion
-		)[0].server_setting_file_format;
-		return format;
 	};
 
 	/**
@@ -280,7 +202,7 @@
 
 		<form bind:this={formElement} on:submit|preventDefault={generateSettingFile}>
 			<div class="grid gap-4 sm:grid-cols-2 sm:gap-6">
-				{#each selectedVersionSettings as setting}
+				{#each selectedSettings as setting}
 					{#if setting.type === 'planetext'}
 						<PlanetextSetting {setting} {forceEnableDisabledItems} bind:formValue={formValues[setting.key]} />
 					{:else if setting.type === 'string'}
@@ -294,24 +216,7 @@
 					{/if}
 				{/each}
 
-				<div class="sm:col-span-2">
-					<Label for="server-setting-text m-2">設定ファイルテキスト</Label>
-					{#if Object.values(formValues).filter(value => value.allow_empty === false && value.value === '').length > 0 || 
-						Object.values(formValues).filter(value => value.max != null && Number(value.value) > value.max).length > 0 || 
-						Object.values(formValues).filter(value => value.min != null && value.min > Number(value.value)).length > 0}
-						<Alert color="yellow" border>
-							<InfoCircleSolid slot="icon" class="w-4 h-4" />
-							設定値が不正です。必須項目の確認と入力可能な範囲を超えた数値がないか確認してください。
-						</Alert>	
-					{:else}
-						<Textarea
-							id="server-setting-text"
-							rows="15"
-							bind:value={serverSettingFileText}
-							readonly
-							/>	
-					{/if}
-				</div>
+				<SettingTextPreview {selectedServerVersion} {selectedSettings} bind:settingText={settingText} {formValues} />
 
 				<div class="sm:col-span-2">
 					<Button type="submit" class="w-full">設定ファイルを作成</Button>
